Recurso de patch adicionado.

+++ src/main/java/com/supermarket/mvcsupermarket/Repository/EmployeeRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
+public interface EmployeeRepository extends JpaRepository<Employee, Long> {
     Employee findById(long id);
 
-    List<Employee> findByNomeContainingIgnoreCaseOrId(String nome, Integer id);
+    List<Employee> findByNomeContainingIgnoreCaseOrId(String nome, long id);
 }
